Allow a built-in clear command to be included with the rest of the picocli commands
